+++ codegen/src/main/java/io/reactiverse/es4x/codegen/generator/IndexJS.java
@@ -24,6 +24,7 @@ import java.util.HashSet;
 import java.util.Map;
 
 import static io.reactiverse.es4x.codegen.generator.Util.getNPMScope;
+import static io.reactiverse.es4x.codegen.generator.Util.includeFileIfPresent;
 
 public class IndexJS extends Generator<ClassModel> {
 
@@ -50,11 +51,14 @@ public class IndexJS extends Generator<ClassModel> {
     if (index == 0) {
       Util.generateLicense(writer);
 
-      writer.printf("/// <reference types=\"%s\" />\n", getNPMScope(model.getType().getRaw().getModule()));
+      writer.printf("/// <reference types=\"%s\" />\n\n", getNPMScope(model.getType().getRaw().getModule()));
+      writer.printf(
+        "/**\n" +
+          " * @typedef { import(\"es4x\") } Java\n" +
+          " */\n");
       writer.print("module.exports = {\n");
-      if (model.getType().getRaw().getModuleName().equals("vertx")) {
-        writer.print("  AsyncResult: Java.type('io.vertx.core.AsyncResult'),\n");
-      }
+      // include a file if present
+      writer.print(includeFileIfPresent("index.include.js"));
     }
 
     writer.printf("  %s: Java.type('%s')", model.getType().getRaw().getSimpleName(), model.getType().getName());