+++ es4x/src/main/java/io/reactiverse/es4x/ECMAEngine.java
@@ -37,6 +37,8 @@ import java.util.logging.ConsoleHandler;
 import java.util.logging.Handler;
 import java.util.regex.Pattern;
 
+import static io.reactiverse.es4x.impl.AsyncError.parseStrackTraceElement;
+
 public final class ECMAEngine {
 
   private static final Logger LOG = LoggerFactory.getLogger(ECMAEngine.class);
@@ -128,7 +130,11 @@ public final class ECMAEngine {
       .targetTypeMapping(
         Value.class,
         Buffer.class,
-        Value::hasMembers,
+        // ensure that the type really matches
+        v -> {
+          final Value meta = v.getMetaObject();
+          return isScriptObject(v) && meta != null && "ArrayBuffer".equals(meta.getMember("className").asString());
+        },
         v -> {
           if (v.hasMember("nioByteBuffer")) {
             return Buffer.buffer(Unpooled.wrappedBuffer(v.getMember("nioByteBuffer").as(ByteBuffer.class)));
@@ -139,9 +145,59 @@ public final class ECMAEngine {
         })
       // Ensure Arrays are exposed as List when the Java API is accepting Object
       .targetTypeMapping(List.class, Object.class, null, v -> v)
+      // map native Error Object to Throwable
+      .targetTypeMapping(
+        Value.class,
+        Throwable.class,
+        // an error has 2 fields: name + message
+        it -> isScriptObject(it) && it.hasMember("name") && it.hasMember("message"),
+        v -> {
+          final String nameField = v.getMember("name").asString();
+          final String messageField = v.getMember("message").asString();
+          // empty message fields usually it JS prints the name field
+          final Throwable t = new Throwable("".equals(messageField) ? nameField : messageField);
+          // the stacktrace for JS is a single string and we need to parse it back to a Java friendly way
+          if (v.hasMember("stack")) {
+            String stack = v.getMember("stack").asString();
+            String[] sel = stack.split("\n");
+            StackTraceElement[] elements = new StackTraceElement[sel.length - 1];
+            for (int i = 1; i < sel.length; i++) {
+              elements[i-1] = parseStrackTraceElement(sel[i]);
+            }
+            t.setStackTrace(elements);
+          }
+
+          return t;
+        })
       .build();
 
     fileSystem = new VertxFileSystem(vertx);
+  }
+
+  /**
+   * Is script like object.
+   *
+   * @param v the value to consider
+   * @return true for non null, unknown shape and not Proxy
+   */
+  private static boolean isScriptObject(Value v) {
+    return
+      // nullability
+      !v.isNull() &&
+      // primitives
+      !v.isBoolean() &&
+      !v.isDate() &&
+      !v.isDuration() &&
+      !v.isInstant() &&
+      !v.isNumber() &&
+      !v.isString() &&
+      !v.isTime() &&
+      !v.isTimeZone() &&
+      // exceptions
+      !v.isException() &&
+      // rest
+      !v.isNativePointer() &&
+      !v.isHostObject();
   }
 
   private void registerCodec(Class className) {
@@ -193,7 +249,9 @@ public final class ECMAEngine {
       .allowPolyglotAccess(polyglotAccess);
 
     // allow specifying the custom ecma version
-    builder.option("js.ecmascript-version", System.getProperty("js.ecmascript-version", "2019"));
+    if (System.getProperty("js.ecmascript-version") != null) {
+      builder.option("js.ecmascript-version", System.getProperty("js.ecmascript-version"));
+    }
 
     // the instance
     final Context context = builder.build();
@@ -201,7 +259,7 @@ public final class ECMAEngine {
     // install the codec if needed
     if (codecInstalled.compareAndSet(false, true)) {
       // register a default codec to allow JSON messages directly from GraalJS to the JVM world
-      final Consumer callback = value -> registerCodec(value.getClass());
+      final Consumer<?> callback = value -> registerCodec(value.getClass());
 
       context.eval(
         Source.newBuilder("js", "(function (fn) { fn({}); })", "<class-lookup>").cached(false).internal(true).buildLiteral()