+++ es4x/src/main/java/io/reactiverse/es4x/impl/future/ES4XFailedFuture.java
@@ -51,7 +51,7 @@ public class ES4XFailedFuture<T> implements Future<T>, Promise<T>, Thenable {
   }
 
   @Override
-  public Future<T> setHandler(Handler<AsyncResult<T>> handler) {
+  public Future<T> onComplete(Handler<AsyncResult<T>> handler) {
     handler.handle(this);
     return this;
   }