+++ examples/react-ssr-basic/src/client/pages/home/home.js
@@ -1,0 +1,11 @@
+import React, { Component, Fragment } from "react";
+
+export default class Home extends Component {
+	render() {
+		return (
+			<Fragment>
+				<p>this is Home</p>
+			</Fragment>
+		);
+	}
+}