+++ examples/react-ssr-basic/src/client/pages/not-found/not-found.js
@@ -1,0 +1,11 @@
+import React, { Component, Fragment } from "react";
+
+export default class NotFound extends Component {
+	render() {
+		return (
+			<Fragment>
+				<p>this is Not Found</p>
+			</Fragment>
+		);
+	}
+}