+++ generator/io.vertx/vertx-core/index.include.d.ts
@@ -1,0 +1,10 @@
+export interface Handler<T> {
+  handle(arg0: T) : void;
+}
+
+export interface AsyncResult<T> {
+  succeeded() : boolean;
+  failed() : boolean;
+  cause() : Throwable | null;
+  result() : T | null;
+}