+++ generator/io.vertx/vertx-core/index.include.js
@@ -1,0 +1,1 @@
+  AsyncResult: Java.type('io.vertx.core.AsyncResult'),