+++ pm/src/main/java/io/reactiverse/es4x/commands/Install.java
@@ -178,8 +178,13 @@ public class Install implements Runnable {
 
     if (force || libs.exists()) {
       final double version = Double.parseDouble(System.getProperty("java.specification.version"));
-      final String vm = System.getProperty("java.vm.name");
-      if (!vm.toLowerCase().contains("graalvm")) {
+      final boolean isGraalVM =
+        System.getProperty("java.vm.name", "").toLowerCase().contains("graalvm") ||
+        // from graal 20.0.0 the vm name doesn't contain graalvm in the name
+        // but it is now part of the vendor version
+        System.getProperty("java.vendor.version", "").toLowerCase().contains("graalvm");
+
+      if (!isGraalVM) {
 
         // not on graal, install graaljs and dependencies
         warn("Installing GraalJS...");